The base.dart file is becoming very hard to manage. Split off time-related things to separate module.
