+++ test/src/base_test.dart
@@ -72,20 +72,6 @@ void main() {
     expect(GpsHistoryException('test').toString(), 'GpsHistoryException: test');
   });
 
-  test('Time comparisons', () {
-    // Simple comparisons on standalone entities.
-    DateTime intToDateTime(int time) {
-      return DateTime.utc(time);
-    }
-
-    expect(compareTime(intToDateTime(1), intToDateTime(2)),
-        TimeComparisonResult.before);
-    expect(compareTime(intToDateTime(1), intToDateTime(1)),
-        TimeComparisonResult.same);
-    expect(compareTime(intToDateTime(2), intToDateTime(1)),
-        TimeComparisonResult.after);
-  });
-
   group('Test GpsPoint', () {
     testBasicPoint(GpsPoint.allZero.copyWith);
     testEqualityOfPoints(GpsPoint.allZero.copyWith);
